+++ changelog.md
@@ -18,6 +18,8 @@ Codion Change Log
 - ControlIcon.controlIcon(icon) factory method added, when all you have is a single icon.
 - ListBoxBuilder.string() renamed formatter().
 - FilterTableColumnModel.ColumnSelection added.
+- FilterTable.copyToClipboard() now only includes selected columns if column selection is enabled.
+- FilterTable, copyCell control now only enabled when a lead selection is present.
 ### is.codion.framework.domain
 - ColumnDefinition.Builder.hasDatabaseDefault() renamed withDefault().
 - Column.ColumnDefiner.column(Function<Column, ColumnDefinition.Builder) added, for shared column definition policies.